requirements.txt
accelerate>=0.30.0
bitsandbytes>=0.43.1
datasets>=2.19.0
peft>=0.11.1
pyyaml>=6.0.1
scipy>=1.13.0
sentencepiece>=0.2.0
tensorboard>=2.14.0
transformers>=4.42.0
trl>=0.9.6
